refactor: modal sub-folder lower case
