+++ frontend/svelte/src/lib/modals/accounts/AddNewAccount.svelte
