+++ frontend/svelte/src/lib/modals/neurons/StakeNeuron.svelte
