+++ frontend/svelte/src/routes/Neurons.svelte
@@ -6,7 +6,7 @@
   import { i18n } from "../lib/stores/i18n";
   import Toolbar from "../lib/components/ui/Toolbar.svelte";
   import NeuronCard from "../lib/components/neurons/NeuronCard.svelte";
-  import CreateNeuronModal from "../lib/modals/CreateNeuronModal/CreateNeuronModal.svelte";
+  import CreateNeuronModal from "../lib/modals/neurons/CreateNeuronModal.svelte";
 
   // TODO: To be removed once this page has been implemented
   onMount(() => {
